get RollingTrend to work, separate util methods into own file
