+++ featuretools/primitives/api.py
@@ -1,4 +1,5 @@
 # flake8: noqa
+from .rolling_primitive_utils import apply_roll_with_offset_gap, roll_series_with_gap
 from .standard import *
 from .utils import (
     get_aggregation_primitives,
