+++ featuretools/primitives/standard/rolling_transform_primitive.py
@@ -4,8 +4,11 @@ from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import Datetime, Double
 
 from featuretools.primitives.base.transform_primitive_base import TransformPrimitive
+from featuretools.primitives.rolling_primitive_utils import (
+    apply_roll_with_offset_gap,
+    roll_series_with_gap,
+)
 from featuretools.utils import calculate_trend
-from featuretools.primitives.utils import RollingPrimitiveUtils
 
 
 class RollingMax(TransformPrimitive):
@@ -100,14 +103,14 @@ class RollingMax(TransformPrimitive):
     def get_function(self):
         def rolling_max(datetime, numeric):
             x = pd.Series(numeric.values, index=datetime.values)
-            rolled_series = RollingPrimitiveUtils.roll_series_with_gap(
+            rolled_series = roll_series_with_gap(
                 x, self.window_length, gap=self.gap, min_periods=self.min_periods
             )
 
             if isinstance(self.gap, str):
                 additional_args = (self.gap, max, self.min_periods)
                 return rolled_series.apply(
-                    RollingPrimitiveUtils.apply_roll_with_offset_gap,
+                    apply_roll_with_offset_gap,
                     args=additional_args,
                 ).values
             return rolled_series.max().values
@@ -206,14 +209,14 @@ class RollingMin(TransformPrimitive):
     def get_function(self):
         def rolling_min(datetime, numeric):
             x = pd.Series(numeric.values, index=datetime.values)
-            rolled_series = RollingPrimitiveUtils.roll_series_with_gap(
+            rolled_series = roll_series_with_gap(
                 x, self.window_length, gap=self.gap, min_periods=self.min_periods
             )
 
             if isinstance(self.gap, str):
                 additional_args = (self.gap, min, self.min_periods)
                 return rolled_series.apply(
-                    RollingPrimitiveUtils.apply_roll_with_offset_gap,
+                    apply_roll_with_offset_gap,
                     args=additional_args,
                 ).values
             return rolled_series.min().values
@@ -305,14 +308,14 @@ class RollingMean(TransformPrimitive):
     def get_function(self):
         def rolling_mean(datetime, numeric):
             x = pd.Series(numeric.values, index=datetime.values)
-            rolled_series = RollingPrimitiveUtils.roll_series_with_gap(
+            rolled_series = roll_series_with_gap(
                 x, self.window_length, gap=self.gap, min_periods=self.min_periods
             )
 
             if isinstance(self.gap, str):
                 additional_args = (self.gap, np.mean, self.min_periods)
                 return rolled_series.apply(
-                    RollingPrimitiveUtils.apply_roll_with_offset_gap,
+                    apply_roll_with_offset_gap,
                     args=additional_args,
                 ).values
             return rolled_series.mean().values
@@ -410,7 +413,7 @@ class RollingSTD(TransformPrimitive):
     def get_function(self):
         def rolling_std(datetime, numeric):
             x = pd.Series(numeric.values, index=datetime.values)
-            rolled_series = RollingPrimitiveUtils.roll_series_with_gap(
+            rolled_series = roll_series_with_gap(
                 x, self.window_length, gap=self.gap, min_periods=self.min_periods
             )
 
@@ -422,7 +425,7 @@ class RollingSTD(TransformPrimitive):
                 additional_args = (self.gap, _pandas_std, self.min_periods)
 
                 return rolled_series.apply(
-                    RollingPrimitiveUtils.apply_roll_with_offset_gap,
+                    apply_roll_with_offset_gap,
                     args=additional_args,
                 ).values
             return rolled_series.std().values
@@ -446,13 +449,13 @@ class RollingTrend(TransformPrimitive):
     def get_function(self):
         def rolling_trend(datetime, numeric):
             x = pd.Series(numeric.values, index=datetime.values)
-            rolled_series = RollingPrimitiveUtils.roll_series_with_gap(
+            rolled_series = roll_series_with_gap(
                 x, self.window_length, gap=self.gap, min_periods=self.min_periods
             )
             if isinstance(self.gap, str):
                 additional_args = (self.gap, np.mean, self.min_periods)
                 return rolled_series.apply(
-                    RollingPrimitiveUtils.apply_roll_with_offset_gap,
+                    apply_roll_with_offset_gap,
                     args=additional_args,
                 ).values
             return rolled_series.apply(calculate_trend).values
@@ -548,16 +551,16 @@ class RollingCount(TransformPrimitive):
     def get_function(self):
         def rolling_count(datetime):
             x = pd.Series(1, index=datetime)
-            rolled_series = RollingPrimitiveUtils.roll_series_with_gap(
+            rolled_series = roll_series_with_gap(
                 x, self.window_length, gap=self.gap, min_periods=self.min_periods
             )
 
             if isinstance(self.gap, str):
-                # Since _apply_roll_with_offset_gap doesn't artificially add nans before rolling,
+                # Since apply_roll_with_offset_gap doesn't artificially add nans before rolling,
                 # it produces correct results
                 additional_args = (self.gap, len, self.min_periods)
                 return rolled_series.apply(
-                    RollingPrimitiveUtils.apply_roll_with_offset_gap,
+                    apply_roll_with_offset_gap,
                     args=additional_args,
                 ).values
